// packages/core/src/entity/preference/enums.ts
export enum Side {
  Giving = 'giving',
  Receiving = 'receiving',
  Neutral = 'neutral',
}

export enum PreferenceLevel {
  Light = 'light',
  Medium = 'medium',
  Extreme = 'extreme',
}

/**
 * values in percentage
 */
export enum Likeness {
  Dislike = 0,
  Ehh = 25,
  Maybe = 40,
  Like = 60,
  Love = 75,
  Favorite = 90,
}